parse content length as usize, not u64
This addresses a [build
failure](https://github.com/scottlamb/rtsp-types/pull/1/checks?check_run_id=2651930731) on 32-bit ARM:

```
   Compiling rtsp-types v0.0.1 (/home/runner/work/rtsp-types/rtsp-types)
error[E0277]: the trait bound `u64: ToUsize` is not satisfied
   --> src/parser.rs:161:23
    |
161 |     move |input| take(content_length)(input)
    |                       ^^^^^^^^^^^^^^ the trait `ToUsize` is not implemented for `u64`
    |
   ::: /home/runner/.cargo/registry/src/github.com-1ecc6299db9ec823/nom-6.1.2/src/bytes/streaming.rs:426:6
    |
426 |   C: ToUsize,
    |      ------- required by this bound in `nom::bytes::streaming::take`

error: aborting due to previous error

For more information about this error, try `rustc --explain E0277`.
error: could not compile `rtsp-types`
```

+++ src/parser.rs
@@ -155,7 +155,7 @@ fn body(headers: &[HeaderRef]) -> impl for<'a> Fn(&'a [u8]) -> IResult<&'a [u8],
     let content_length = headers
         .iter()
         .find(|h| &h.name.to_ascii_uppercase() == "CONTENT-LENGTH")
-        .map(|h| str::parse::<u64>(&h.value).unwrap())
+        .map(|h| str::parse::<usize>(&h.value).unwrap())
         .unwrap_or(0);
 
     move |input| take(content_length)(input)
